pyproject.toml
[project]
name = "pyautox"
version = "0.1.0"
description = "Modern macOS GUI automation library for Python"
authors = [{name = "PyAutoX Contributors"}]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "BSD-3-Clause"}
keywords = ["automation", "gui", "macos", "pyautogui", "mouse", "keyboard", "screenshot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pyobjc-core>=10.0; sys_platform == 'darwin'",
    "pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'",
    "pyobjc-framework-ApplicationServices>=10.0; sys_platform == 'darwin'",
]

[project.optional-dependencies]
locate = [
    "opencv-python>=4.8.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/VoyagerTechHQ/pyautox"
Documentation = "https://github.com/VoyagerTechHQ/pyautox#readme"
Repository = "https://github.com/VoyagerTechHQ/pyautox"
Issues = "https://github.com/VoyagerTechHQ/pyautox/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pyautox"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313", "py314"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["N802", "N803"]  # camelCase names for PyAutoGUI API compatibility
